Set properly HasAuthorHighlightColors flag for links
Right now that flag was set for ComputedStyle of both visited
and unvisited links, even when only of them have styles set.

This patch adds a new flag "is_visited_highlight_colors"
so we can differentiate both cases, and only set
HasAuthorHighlightColors when appropriated.

This patch fixes the issue when HighlightInheritance flag
is enabled.

Bug: 1314629
Test: external/wpt/css/css-pseudo/selection-link-001.html
Test: external/wpt/css/css-pseudo/selection-link-002.html
Test: external/wpt/css/css-pseudo/selection-link-003.html
Change-Id: I0ee422bb7eb2098774dc829512f17d3d228f9556