Reduce ramp-up time and re-enable simulcast tests.
negotiateSimulcastAndWaitForVideo() is updated to send 360p as the top
layer instead of 720p, this is believed to improve ramp-up time and
thus reduce the risk of bots randomly timing out.

This CL speculatively re-enables the tests on all bots, but note that
we've never been able to reproduce the timeout locally (either before or
after this CL).

Bug: 1230534
Change-Id: If9b540f87f1e39bda140f89fa04e55555b7e85d4